+++ src/main/kotlin/com/codescene/jetbrains/services/telemetry/TelemetryService.kt
@@ -2,6 +2,7 @@ package com.codescene.jetbrains.services.telemetry
 
 import com.codescene.ExtensionAPI
 import com.codescene.data.telemetry.TelemetryEvent
+import com.codescene.jetbrains.config.global.CodeSceneGlobalSettingsStore
 import com.codescene.jetbrains.services.BaseService
 import com.codescene.jetbrains.util.Constants
 import com.codescene.jetbrains.util.Log
@@ -25,7 +26,11 @@ class TelemetryService(): BaseService(), Disposable {
         fun getInstance(): TelemetryService = service<TelemetryService>()
     }
 
-    fun logUsage(eventName: String, eventData: Map<String, Any> = mutableMapOf<String, Any>()) {
+    fun logUsage(eventName: String, eventData: Map<String, Any> = mutableMapOf()) {
+        val isTelemetryEnabled = CodeSceneGlobalSettingsStore.getInstance().state.telemetryConsentGiven
+
+        if (!isTelemetryEnabled) return
+
         val extendedName = "${Constants.TELEMETRY_EDITOR_TYPE}/$eventName"
         // TODO: Get user ID of logged in user when authentication is implemented
         val userId: String? = null