+++ src/main/kotlin/com/codescene/jetbrains/util/Constants.kt
@@ -15,11 +15,19 @@ object Constants {
     const val IMAGES_BASE_PATH = "images/"
     const val LOGO_PATH = "${IMAGES_BASE_PATH}logo.svg"
     const val CODESCENE_URL = "https://codescene.com"
-    const val CODE_HEALTH_URL = "$CODESCENE_URL/product/code-health#:~:text=Code%20Health%20is%20an%20aggregated,negative%20outcomes%20for%20your%20project"
+    const val CODE_HEALTH_URL =
+        "$CODESCENE_URL/product/code-health#:~:text=Code%20Health%20is%20an%20aggregated,negative%20outcomes%20for%20your%20project"
     const val TERMS_AND_CONDITIONS_URL = "$CODESCENE_URL/policies"
     const val AI_PRINCIPLES_URL = "$CODESCENE_URL/ace/principles"
     const val CONTACT_URL = "$CODESCENE_URL/company/contact-us"
     const val SUPPORT_URL = "https://supporthub.codescene.com/kb-tickets/new"
+
+    const val REPOSITORY_URL = "https://github.com/codescene-oss/codescene-jetbrains"
+    const val TELEMETRY_SAMPLES_URL =
+        "$REPOSITORY_URL/tree/master/src/main/resources/telemetrySamples"
+    const val TELEMETRY_EVENTS_URL =
+        "$REPOSITORY_URL/tree/master/src/main/kotlin/com/codescene/jetbrains/util/TelemetryEvents.kt"
+
     val RED = JBColor(Color(224, 82, 92), Color(224, 82, 92))
     val GREEN = JBColor(Color(79, 159, 120), Color(79, 159, 120))
     val ORANGE = JBColor(Color(250, 163, 125), Color(238, 147, 107))